requirements.txt
requests
jsonschema
pytest
selene==2.0.0rc9
allure-pytest
selenium==4.33.0
webdriver-manager>=4.0.0
python-dotenv